Move `sortKeys` into a helper function.
